+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Changed
 - **Breaking Change** Updated minimum version of `ReactiveSwift` to 3.0.
 [#1470](https://github.com/Moya/Moya/pull/1470) by [@larryonoff](https://github.com/larryonoff).
+- **Breaking Change** Changed the `validate` property of `TargetType` to use new `ValidationType` enum representing valid status codes. [#1505](https://github.com/Moya/Moya/pull/1505) by [@SD10](https://github.com/sd10), [@amaurydavid](https://github.com/amaurydavid). 
 
 # [10.0.1] - 2017-11-23
 ### Fixed